backend/requirements.txt
Django==5.0.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
python-dotenv==1.0.1
django-filter==24.2
drf-spectacular==0.27.1
drf-spectacular-sidecar==2024.7.1
Pillow==10.2.0